mlcloud/models/adapters/yolo.py
"""YOLO11 adapter."""

from typing import Dict, Any
from PIL import Image

from mlcloud.models.registry import ModelMetadata


class YOLO11Adapter:
    """Adapter for YOLO11 models (segmentation and oriented bounding boxes)."""

    def __init__(self, metadata: ModelMetadata):
        """Initialize YOLO11 adapter.
        
        Args:
            metadata: Model metadata
        """
        self.metadata = metadata

    def load(self) -> Any:
        """Load YOLO11 model.
        
        Returns:
            Loaded model
        """
        # Placeholder - full implementation would use ultralytics package
        # from ultralytics import YOLO
        # return YOLO(self.metadata.weights_file)
        return None

    def predict(self, model: Any, image: Image.Image, **kwargs: Any) -> Dict[str, Any]:
        """Run YOLO11 inference.
        
        Args:
            model: Loaded model
            image: Input image
            **kwargs: Additional parameters
            
        Returns:
            Prediction results
        """
        # Placeholder - full implementation would run YOLO11 inference
        return {
            "boxes": [],
            "masks": [],
            "scores": [],
        }

