pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mlcloud"
version = "1.0.0"
description = "A zero-install, platform-agnostic Python CLI that turns any laptop into an on-demand GPU computer-vision workstation with zero vendor lock-in."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "mlcloud contributors"}
]
keywords = ["cvat", "computer-vision", "annotation", "gpu", "aws", "coreweave", "runpod"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "keyring>=24.0.0",
    "requests>=2.31.0",
    "boto3>=1.28.0",
    "boto3-stubs[essential]>=1.28.0",
    "kubernetes>=28.0.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=41.0.0",
    "jinja2>=3.1.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
]

[project.scripts]
mlcloud = "mlcloud.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/mlcloud"
Documentation = "https://github.com/your-org/mlcloud#readme"
Repository = "https://github.com/your-org/mlcloud"
Issues = "https://github.com/your-org/mlcloud/issues"

[tool.hatch.build.targets.wheel]
packages = ["mlcloud"]

[tool.hatch.build.targets.sdist]
include = [
    "/mlcloud",
    "/tests",
    "/examples",
    "/docs",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]
exclude = [
    "/legacy-infrastructure",
    "/htmlcov",
    "/.github",
    "/.pytest_cache",
    "/docs/project",  # Project docs not needed in package
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=mlcloud",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

